requirements.txt
Pyrogram==2.0.106
python-dotenv==0.21.0
sentry-sdk==2.48.0